fix PULSEDEV-10540 pdb: Double serialization in event storage for blob fields
Changed handling of BLOB values so that byte arrays are not serialized. Up to now, BLOB values are always stored as java serialized values; with this change, BLOB values will have a prefix (encoder id) that indicates how the value is encoded followed by the actual value. Two encodings are supported, java serialization (the old encoding) plus null serialization (that from now on will be used for byte arrays).

In order to have backward compatibility, the java serializer encoding id is 0xac, which is a prefix used by the java serialization protocol (0xac is the first byte of STREAM_MAGIC). In addition, on this encoding, the encoding id is considered part of the encoded value.

Existing unit tests already cover scenarios for BLOB byte[] values and values of other types. Added only scenario for BLOB null values.