Adds a lock to ensure batch flush is thread-safe.
Before this fix, 2 threads calling flush concurrently on the same batch could both try to create
overlapping transactions on the same DatabaseEngine. This is problematic and could lead to errors
like 'Cannot commit when autoCommit is enabled'. With this fix only one thread at a time will be
able to enter the transactional section of the flush method.

This fix also fix a problem that would occur if a transaction properly committed but the
de.isTransactionActive() in the finally clause threw an exception, onFlushFailure would still be
called with the temp buffer. So we were notifying of a flush failure that actually succedded.