ref PULSEDEV-16501 pdb: Improve the performance of writing to Oracle CLOB columns
The usage of PreparedStatement#setClob in Oracle 11g was yielding overheads of 3x
to 1000x (depending on the batch size) when compared to using plain
PreparedStatement#setObject for columns backed by VARCHAR2.

With this change we downcast the PreparedStatement to OraclePreparedStatement and
use its setStringForClob method that yields equivalent performance to the VARCHAR2
case for Strings whose length is 32KB or less.