Improve feedback for skipping schema create/update
If a user tries to create/update an entity, and if the schema policy is
set to `none`, nothing will happen. This patch adds a warn entry to
notify that the operation was skipped, thus helping troubleshooting
related issues.

On `updateEntity`, it also only calls `createPreparedStatementForInserts` if
policy is different than `none`. This was throwing an exception on H2.
The reasoning is that if nothing was changed, no prepared statements
need to be updated.