Add indexes when a new column is added to an entity
When a column is being added to a table that already exists, it's indexes
were being left out. This patch fixes that by adding the indexes that
consider the new column(s).

A new `addIndexes` method was added to `AbstractDatabaseEngine` that
creates specific indexes.

Missing on this patch:
* Wasn't able to test this, missing dependencies and couldn't find
  instructions
* Would like a unit test for this, but don't have the env and wouldn't
  know how to achieve it for this scenario

Related possible future work:
* Implement drop indexes
* When we remove columns, we should also remove related indexes
  (I guess it would throw an exception otherwise).
* Use the same logic for FKs. ATM we're droping/adding them all when
  adding or removing new fields. That could be very slow on large tables