ref PULSEDEV-18139 pdb: Fix incorrect shutdown in AbstractBatch
This change fixes a small concurrency issue that was occurring in AbstractBatch
when the batch scheduler was scheduling a run while the destroy
method had just been called, leading to an unexpected long wait for the
destroy method to return.
Since those two methods were `synchronized`
the following could happen:
   - Thread A entered `destroy` thus acquiring the implicit lock
   - The scheduler period would expire and a new task would be
   launched but since the `run` method was synchronized the task's
   thread (B)  would not proceed
   - Thread A would now call `shutdownNow` on the ExecutorService and
   await termination.
   - Since Thread B was waiting on the implicit lock to enter `run` the
     `destroy` code would wait until the timeout expired before
     returning.

There was reason to keep the methods synchronized as synchronization was
already being performed at a different level.