ref PULSEDEV-19649 pdb: Add DB test profiles and docker containers
- Add a Maven test profile for each DB vendor, which starts a docker container
with that vendor's server and runs the tests on it;

- Update dependencies versions;

- Also fix a problem that caused some tests to fail on the first run:
with "drop-create" schema policy, adding a table will involve trying
first to delete that table; if that table doesn't exist and we are inside
a transaction, there will be an SQL error while deleting the non existing
table, causing the current transaction to become "failed"; in this state
subsequent commits will fail.
Affected tests were changed to add the tables first and only then
begin a transaction.