Clear batch data when rolling back a transaction to ensure ACID properties
Currently, a transaction rollback executes a rollback statement on the
database. However PDB might still have prepared statements with batch data that,
unless cleared, can still be saved on the database, thereby violating the ACID
properties. This commit ensures that the prepared statements are cleared when
rolling back a transaction.