Fix connection bug and add test profile for H2 remote
Summary:
H2 remote, for some reason, doesn't support the JDBC methods getSchema()
 and setSchema(String).

This was known in the last commit to make PDB use schmema property, so
 getSchema() got replaced with a query when the connection is H2 remote.
The same was done for setSchema, but the query used a PreparedStatement
 to replace the schema identifier, which is not allowed (PreparedStatements
 replace values only).
This commit replaces the PreparedStatement with a normal Statement.

DB2 is also using a PreparedStatement to set schema, but it works with
 that engine, so nothing was changed there.

This commit fixes the problem in H2 remote and adds a new 'h2remote' profile.