[Oracle] Batch inserts/updates with LOBs don't free temporary LOB (#114)
On batch updates with LOB columns with a size greater than 4000 bytes, the oracle driver uses a temporary LOB that is never freed because PDB uses a single PreparedStatement for the whole session. This commit avoids this resource leak by using explicit LOBs on these circumstances, and releasing them explicitly when the batch is flushed. The OracleEngine was modified by adding an attribute with the actions that must be executed after the flush of the current batch, and ensuring that actions are added on addBatch() and performed on flush(). The state of this attribute reflects what was added to the batch of the current insert PreparedStatement, so it has the same resource release requirements as that PreparedStatement.