Adds support for timeouts in select queries (#123)
This provides a mechanism to specify a default timeout for all select queries created by a given
DatabaseEngine plus the possibility to override the configured value on individual queries. It
also exposes the Statement#cancel() method, that allows a thread to cancel a query running on other
query and is supported by the production-ready JDBC drivers supported by pdb.