Add retryable flag to DatabaseEngineRuntimeException
Summary:
When a transaction fails on commit, because a concurrent transaction
interfered with it (server was unable to serialize, or deadlock was
detected), the server sends back information indicating the error.

These errors can be retried (and will probably succeed then), so this
commit adds the information that something can be retried in a flag of
DatabaseEngineRuntimeException, so that the client application knows
that it can retry the transaction.