Change "allowColumnDrop" default to false
Summary:
Having the default at true, although convenient to keep the DB tables
synchronized with the respective PDB entities, may lead to unintended
loss of data (by default, if a column disappears from an entity, PDB
will update the table by dropping the respective column along with the
historical data in it.

For safety, this commit changes that default to false.