Add support for CockroachDB (#132)
Summary:
CockroachDB uses the same driver as PostgreSQL, but there are some
incompatibilities, and a new profile needed to be setup for tests to
allow using a CockroachDB Docker image.

This commit adds a new CockroachDB engine that extends the one for
PostgreSQL to address the incompatibilities and makes some small
adjustments to all code accordingly.
A new "cockroach" profile was added to the pom to run the tests with a
CockroachDB instance.

The pom was changed by adding new "image.<profile_name>" flags, one for
each profile, that allows specifying docker images different from the
defaults (e.g. new versions) to run the tests with.
Those defaults are the images/versions that were already in the pom.