Cleanups in preparation for more retryable exceptions
Summary:
Retryable errors when running queries may occur in other places
besides the commit.

In preparation for that, the method
 `persist(final String name, final EntityEntry entry)`
was refactored to AbstractDatabaseEngine (it was implemented in
all engines, but the code was always the same).

This way, any try-catch for the retryable exceptions only needs
to be done in the persist with autoInc parameter.

While at it, the method `translateType(DbColumn c)` was also
repeated in all engines and was moved to AbstractDatabaseEngine.

Other cleanups highlighted by the IDE were also done.