Fix computed column boolean conversion
In MySql and SqlServer, the result column  failed
when called on a computed column of boolean type. This is caused
by the implicit type conversions that the driver does when fetching
a value. For regular columns, the type is well known and the
conversion between 0/1 to false/true works. For computed columns,
the type conversions fail because the driver doesn't know the type.

See the new unit-test a simple scenario that failed. Also see the
new test Javadoc for more details on another (failed) approach.