Add support for CAST to MySQL
Summary:
MySQL was left out when CAST was introduced.
This is not a desired situation, because PDB aims to present
features that work across all supported DB engines seamlessly.
Since MySQL does in fact support CAST, this commit adds it to
that engine.
The tests for CAST were also improved.