Splits values in IN/NOTIN expressions for Oracle
Oracle database can only handle a maximum of 1000 values in IN clauses (there is no such limit for IN SELECT).
A query containing an IN clause with more than 1000 values will result in an error.
This commit splits such IN clauses into equivalent expressions, with multiple IN clauses with 1000 values or less connected by ORs (or connected by ANDs for when NOT IN is used).