Add mapping for unique constraint violation exception across engines
Summary:
This fix adds a new exception, DatabaseEngineUniqueConstraintViolationException, to map a unique constraint violation in the database. This exeption is thrown when a statement attempts to insert a record with a key that already exists in the table.

Some Exception Handlers needed to be added to map specific error codes/sql states for that engine.

Why it was introduced:
When a batch with duplicated entries tried to be inserted, it failed due to a unique constraint violation.
A generic DatabaseEngineException was thrown and thus, it was difficult to access the cause of the exception without having a specific exception mapping for that cause.

Testing:
- A new test was added that verifies the exception when we have a batch with duplicated entries and a primary key constraint on the table.
- The test suite was run for every supported engine.