Add ASCII function (#377)
Summary:
Adds the ASCII string function.
This commit also adds STDDEV_POP (standard deviation, but for the population
instead of a sample), and improves translation of functions in general.
The new functions are tested in EngineGeneralTest, where some tests that
performed the same actions on related functions were merged, to reduce the
number of lines.