Try to reconnect at max 3x when connection is down in `AbstractDatabaseEngine#getPSIterator`
It was found out that sometimes the connection to the Database could be lost
when calling the method AbstractDatabaseEngine#createResultIterator.
In order to make it reconnect and redo the desired query, this commit makes the
method AbstractDatabaseEngine#getPSIterator call the
`AbstractDatabaseEngine#createResultIterator` at maximum 3 times, where in each
attempt, it reconnects to the database first.
By doing this, it avoids the PDB user to having to manually reconnect the
connection and redo all the `PreparedStatements` as reported in the exception
that was being thrown "Connection was lost and restablished. You need to reset
the prepared statement parameters and re-execute the statement."