Reverts the Support to Upsert/Insert Ignore on PDB (#471)
Recently, it was added a support for the upsert of entries in the pdb library.

The feature added supported for entities to, when applicable, be able to be upserted or depending on the engine implementation, an insert ignore or even merge. The feature also enabled the upsert feature for batch processing, i.e. entries could now use the new upsert API. However, after its usage it was possible to conclude it decreased performance when client application used this new API. Giving the degrading performance and also the log messages some of the entities generated for not having the upsert statement available, it was decided to remove this feature entirely from the pdb code-base.

---
* Revert "Removes error logging for UPSERT command when creating an entity (#407)"
This reverts commit 94ba1a540458a4e3fbc165921506d18cef103945.

* Revert "Add Support to Upsert/Insert Ignore on PDB (#392)"
This reverts commit 46f29b2b5471e03348040dd143dddffd70c3adf6.

* Revert "Add Support to Upsert/Insert Ignore on PDB (#389)"
This reverts commit 963cd305a0c442b5f8786eed8d4350561d1d4336.

* Revert "Add Support to Upsert/Insert Ignore on PDB (#388)"
This reverts commit d4f7c4e0be586bfb77839fec8ddd143697950eb3.

* Revert "Add Support to Upsert/Insert Ignore on PDB"
This reverts commit d626297e911d4770e18d5cd19c8260edbe70d8a5.

* Revert "Add Support to Upsert/Insert Ignore on PDB"
This reverts commit 1d32f7ff97e064a9c26f4c6aedab2d0b66b3f3dd.

* Revert "Add Support to Upsert/Insert Ignore on PDB"
This reverts commit 36a434c8f10bb548baccdb7e05f9979673277653.

* Revert "Add Support to Upsert/Insert Ignore on PDB"
This reverts commit 1658965c76d8a87623f3b547681fd3414050504a.

* Revert "Add Support to Upsert/Insert Ignore on PDB"
This reverts commit d2f153ba0430f77cdad015f9cb4b027bce64169f.

* Revert "Add Support to Upsert/Insert Ignore on PDB"
This reverts commit 2f0d1ffa

* Revert "Add Support to Upsert/Insert Ignore on PDB"
This reverts commit 348a22dd

* Revert "Add Support to Upsert/Insert Ignore on PDB"
This reverts commit e100a516